+++ Front/components/MapComponet/index.js
@@ -1,34 +1,6 @@
-import React, { useEffect, useState } from "react";
+import React from "react";
 import { StyledMapContainer } from "./styles";
-
-/* Marker 스타일링 */
-const Marker = ({ lat, lng }) => {
-  const convertLatLonToScreenCoords = (latitude, longitude) => ({
-    x: Math.min(Math.max(latitude, 0), 100),
-    y: Math.min(Math.max(longitude, 0), 100),
-  });
-
-  const [position, setPosition] = useState({ x: 0, y: 0 });
-
-  useEffect(() => {
-    const { x, y } = convertLatLonToScreenCoords(lat, lng);
-    setPosition({ x, y });
-  }, [lat, lng]); //lat 또는 lng가 변경될 때마다 실행
-  return (
-    <div
-      style={{
-        position: "absolute",
-        left: `${lat}%`,
-        top: `${lng}%`,
-        transform: `translate(-50%, -50%)`,
-        fontSize: "2rem",
-        transition: "left 0.5s linear, top 0.5s linear",
-      }}
-    >
-      📍
-    </div>
-  );
-};
+import Marker from "../Marker";
 
 const MapComponet = ({ turtlebots }) => {
   return (
