+++ Front/components/Marker/style.js
@@ -1,0 +1,8 @@
+import styled from "@emotion/styled";
+
+export const StyledMarker = styled.div`
+  position: absolute;
+  // transform: translate(0%, 0%);
+  fontsize: 2rem;
+  transition: left 0.5s linear, top 0.5s linear;
+`;
