+++ README.md
@@ -4,6 +4,13 @@
 
 1. 라이브러리 패키지 관리는 yarn.lock으로 진행
 
+### TO-DO
+
+1. FLASK 통신 (웹소켓 라이브러리)
+2. FLASK에서 api 받아와서 Marker 동작 확인하기
+3. 프리셋 또는 Marker를 움직이는 기능 만들기
+4. 이외 기능 추가하기
+
 # BACK-END
 
 # ROS
